+++ tensorflow_io/core/python/api/experimental/__init__.py
@@ -24,7 +24,6 @@ from tensorflow_io.core.python.api.experimental import image
 from tensorflow_io.core.python.api.experimental import text
 from tensorflow_io.core.python.api.experimental import columnar
 from tensorflow_io.core.python.api.experimental import color
-from tensorflow_io.core.python.api.experimental import audio
 from tensorflow_io.core.python.api.experimental import streaming
 from tensorflow_io.core.python.api.experimental import filter
 from tensorflow_io.core.python.api.experimental import elasticsearch
